Split functionality for generating docs into separate .py file
